+++ web/packages/api/src/forKusama.ts
@@ -22,7 +22,8 @@ import {
     buildTransferKusamaToPolkadotExportXCM,
     buildTransferPolkadotToKusamaExportXCM,
 } from "./xcmBuilderKusama"
-import { Asset, AssetRegistry, Parachain, getAssetHubConversationPalletSwap } from "./assets_v2"
+import { getAssetHubConversationPalletSwap } from "./assets_v2"
+import { Asset, AssetRegistry, Parachain } from "./types"
 import {
     CallDryRunEffects,
     EventRecord,
