+++ web/packages/api/src/parachains/acala.ts
@@ -1,4 +1,4 @@
-import { AssetMap, PNAMap } from "../assets_v2"
+import { AssetMap, PNAMap } from "../types"
 import { ParachainBase } from "./parachainBase"
 import { convertToXcmV3X1, getTokenFromLocation } from "../xcmBuilder"
