+++ web/packages/api/src/parachains/bifrost.ts
@@ -1,4 +1,4 @@
-import { AssetMap, PNAMap } from "../assets_v2"
+import { AssetMap, PNAMap } from "../types"
 import { ParachainBase } from "./parachainBase"
 import { DOT_LOCATION, getTokenFromLocation } from "../xcmBuilder"
