+++ web/packages/api/src/parachains/moonbeam.ts
@@ -1,7 +1,7 @@
 import { ApiPromise } from "@polkadot/api"
 import { ethers } from "ethers"
 import { ParachainBase } from "./parachainBase"
-import { AssetMap, PNAMap } from "../assets_v2"
+import { AssetMap, PNAMap } from "../types"
 import { convertToXcmV3X1, DOT_LOCATION, getTokenFromLocation } from "../xcmBuilder"
 
 const MOONBEAM_ERC20_ABI = [
