+++ web/packages/api/src/toEthereum_v2.ts
@@ -1,7 +1,7 @@
 import { ApiPromise } from "@polkadot/api"
 import { AddressOrPair, SignerOptions, SubmittableExtrinsic } from "@polkadot/api/types"
 import { Codec, ISubmittableResult } from "@polkadot/types/types"
-import { BN, hexToU8a, isHex, numberToHex, stringToU8a, u8aToHex } from "@polkadot/util"
+import { BN, hexToU8a, isHex, stringToU8a, u8aToHex } from "@polkadot/util"
 import { blake2AsHex, decodeAddress, xxhashAsHex } from "@polkadot/util-crypto"
 import {
     bridgeLocation,
@@ -18,15 +18,8 @@ import {
     buildExportXcmForERC20,
     HERE_LOCATION,
 } from "./xcmBuilder"
-import {
-    Asset,
-    AssetRegistry,
-    ERC20Metadata,
-    EthereumChain,
-    getAssetHubConversationPalletSwap,
-    padFeeByPercentage,
-    Parachain,
-} from "./assets_v2"
+import { getAssetHubConversationPalletSwap, padFeeByPercentage } from "./assets_v2"
+import { Asset, AssetRegistry, ERC20Metadata, Parachain } from "./types"
 import { getOperatingStatus, OperationStatus } from "./status"
 import { IGatewayV1 as IGateway } from "@snowbridge/contract-types"
 import {
@@ -36,55 +29,8 @@ import {
     XcmDryRunEffects,
 } from "@polkadot/types/interfaces"
 import { Result } from "@polkadot/types"
-import {
-    AbstractProvider,
-    Contract,
-    ContractTransaction,
-    FeeData,
-    TransactionReceipt,
-} from "ethers"
+import { FeeData } from "ethers"
 import { paraImplementation } from "./parachains"
-
-const PALLET_XCM_PRECOMPILE = [
-    {
-        inputs: [
-            {
-                components: [
-                    { internalType: "uint8", name: "parents", type: "uint8" },
-                    { internalType: "bytes[]", name: "interior", type: "bytes[]" },
-                ],
-                internalType: "struct XCM.Location",
-                name: "dest",
-                type: "tuple",
-            },
-            {
-                components: [
-                    { internalType: "address", name: "asset", type: "address" },
-                    { internalType: "uint256", name: "amount", type: "uint256" },
-                ],
-                internalType: "struct XCM.AssetAddressInfo[]",
-                name: "assets",
-                type: "tuple[]",
-            },
-            {
-                internalType: "enum XCM.TransferType",
-                name: "assetsTransferType",
-                type: "uint8",
-            },
-            { internalType: "uint8", name: "remoteFeesIdIndex", type: "uint8" },
-            {
-                internalType: "enum XCM.TransferType",
-                name: "feesTransferType",
-                type: "uint8",
-            },
-            { internalType: "bytes", name: "customXcmOnDest", type: "bytes" },
-        ],
-        name: "transferAssetsUsingTypeAndThenAddress",
-        outputs: [],
-        stateMutability: "nonpayable",
-        type: "function",
-    },
-]
 
 export type Transfer = {
     input: {
@@ -105,30 +51,6 @@ export type Transfer = {
         messageId?: string
     }
     tx: SubmittableExtrinsic<"promise", ISubmittableResult>
-}
-
-export type TransferEvm = {
-    input: {
-        registry: AssetRegistry
-        sourceAccount: string
-        beneficiaryAccount: any
-        tokenAddress: string
-        amount: bigint
-        fee: DeliveryFee
-    }
-    computed: {
-        sourceParaId: number
-        sourceAccountHex: string
-        tokenErcMetadata: ERC20Metadata
-        ahAssetMetadata: Asset
-        sourceAssetMetadata: Asset
-        sourceParachain: Parachain
-        messageId: string
-        ethChain?: EthereumChain
-        customXcmHex?: string
-        xcTokenAddress?: string
-    }
-    tx: ContractTransaction
 }
 
 export type DeliveryFee = {
@@ -248,117 +170,6 @@ export async function createTransfer(
             ahAssetMetadata,
             sourceAssetMetadata,
             messageId,
-        },
-        tx,
-    }
-}
-
-export async function createTransferEvm(
-    parachain: ApiPromise,
-    registry: AssetRegistry,
-    sourceAccount: string,
-    beneficiaryAccount: string,
-    tokenAddress: string,
-    amount: bigint,
-    fee: DeliveryFee
-): Promise<TransferEvm> {
-    const { ethChainId, assetHubParaId } = registry
-
-    let sourceAccountHex = sourceAccount
-    if (!isHex(sourceAccountHex)) {
-        sourceAccountHex = u8aToHex(decodeAddress(sourceAccount))
-    }
-    if (sourceAccountHex.length !== 42) {
-        throw Error(`Source address ${sourceAccountHex} is not a 20 byte address.`)
-    }
-
-    const sourceParachainImpl = await paraImplementation(parachain)
-    const { tokenErcMetadata, sourceParachain, ahAssetMetadata, sourceAssetMetadata } =
-        resolveInputs(registry, tokenAddress, sourceParachainImpl.parachainId)
-    if (!sourceParachain.info.evmChainId) {
-        throw Error(`Parachain ${sourceParachainImpl.parachainId} is not an EVM chain.`)
-    }
-    if (!sourceParachain.xcDOT) {
-        throw Error(`Parachain ${sourceParachainImpl.parachainId} does not support XC20 DOT.`)
-    }
-    const ethChain = registry.ethereumChains[sourceParachain.info.evmChainId.toString()]
-    if (!ethChain) {
-        throw Error(
-            `Cannot find eth chain ${sourceParachain.info.evmChainId} for parachain ${sourceParachainImpl.parachainId}.`
-        )
-    }
-    if (!ethChain.precompile) {
-        throw Error(`No precompile for eth chain ${sourceParachain.info.evmChainId}.`)
-    }
-    if (!ethChain.xcDOT) {
-        throw Error(`No XC20 DOT for eth chain ${sourceParachain.info.evmChainId}.`)
-    }
-    if (!ethChain.xcTokenMap || !ethChain.xcTokenMap[tokenAddress]) {
-        throw Error(`No XC20 token for token address ${tokenAddress}.`)
-    }
-
-    const xcTokenAddress = ethChain.xcTokenMap[tokenAddress]
-    const contract = new Contract(ethChain.precompile, PALLET_XCM_PRECOMPILE)
-
-    const messageId = await buildMessageId(
-        parachain,
-        sourceParachainImpl.parachainId,
-        sourceAccountHex,
-        tokenAddress,
-        beneficiaryAccount,
-        amount
-    )
-    const customXcm = buildAssetHubERC20TransferFromParachain(
-        parachain.registry,
-        ethChainId,
-        sourceAccount,
-        beneficiaryAccount,
-        tokenAddress,
-        messageId,
-        sourceParachainImpl.parachainId,
-        fee.returnToSenderExecutionFeeDOT,
-        DOT_LOCATION // TODO: Support Native fee for EVM chains
-    )
-
-    const tx = await contract[
-        "transferAssetsUsingTypeAndThenAddress((uint8,bytes[]),(address,uint256)[],uint8,uint8,uint8,bytes)"
-    ].populateTransaction(
-        // This represents (1,X1(Parachain(1000)))
-        [1, ["0x00" + numberToHex(assetHubParaId, 32).slice(2)]],
-        // Assets including fee and the ERC20 asset, with fee be the first
-        [
-            [ethChain.xcDOT, fee.totalFeeInDot],
-            [xcTokenAddress, amount],
-        ],
-        // The TransferType corresponding to asset being sent, 2 represents `DestinationReserve`
-        2,
-        // index for the fee
-        0,
-        // The TransferType corresponding to fee asset
-        2,
-        customXcm.toHex()
-    )
-
-    tx.from = sourceAccountHex
-    return {
-        input: {
-            registry,
-            sourceAccount,
-            beneficiaryAccount,
-            tokenAddress,
-            amount,
-            fee,
-        },
-        computed: {
-            sourceParaId: sourceParachainImpl.parachainId,
-            sourceAccountHex,
-            tokenErcMetadata,
-            sourceParachain,
-            ahAssetMetadata,
-            sourceAssetMetadata,
-            messageId,
-            ethChain,
-            xcTokenAddress,
         },
         tx,
     }
@@ -595,21 +406,6 @@ export type ValidationResult = {
         assetHubDryRunError: any
     }
     transfer: Transfer
-}
-
-export type ValidationResultEvm = {
-    logs: ValidationLog[]
-    success: boolean
-    data: {
-        bridgeStatus: OperationStatus
-        nativeBalance: bigint
-        dotBalance?: bigint
-        tokenBalance: bigint
-        feeInfo?: FeeInfo
-        sourceDryRunError: any
-        assetHubDryRunError: any
-    }
-    transfer: TransferEvm
 }
 
 export async function validateTransfer(
@@ -888,257 +684,9 @@ export async function validateTransfer(
     }
 }
 
-export async function validateTransferEvm(
-    connections: {
-        sourceParachain: ApiPromise
-        sourceEthChain: AbstractProvider
-        assetHub: ApiPromise
-        gateway: IGateway
-        bridgeHub: ApiPromise
-    },
-    transfer: TransferEvm
-): Promise<ValidationResultEvm> {
-    const { sourceParachain, gateway, bridgeHub, assetHub, sourceEthChain } = connections
-    const { registry, fee, tokenAddress, amount, beneficiaryAccount } = transfer.input
-    const {
-        sourceAccountHex,
-        sourceParaId,
-        sourceParachain: source,
-        messageId,
-        sourceAssetMetadata,
-    } = transfer.computed
-    const { tx } = transfer
-
-    const sourceParachainImpl = await paraImplementation(sourceParachain)
-    const logs: ValidationLog[] = []
-    let dotBalance: bigint | undefined = undefined
-    if (source.features.hasDotBalance) {
-        dotBalance = await sourceParachainImpl.getDotBalance(sourceAccountHex)
-    }
-    let isNativeBalanceTransfer =
-        sourceAssetMetadata.decimals === source.info.tokenDecimals &&
-        sourceAssetMetadata.symbol == source.info.tokenSymbols
-    const [nativeBalance, tokenBalance] = await Promise.all([
-        sourceParachainImpl.getNativeBalance(sourceAccountHex),
-        sourceParachainImpl.getTokenBalance(sourceAccountHex, registry.ethChainId, tokenAddress),
-    ])
-
-    let nativeBalanceCheckFailed = false
-    if (
-        isNativeBalanceTransfer &&
-        fee.totalFeeInNative &&
-        amount + fee.totalFeeInNative > tokenBalance
-    ) {
-        nativeBalanceCheckFailed = true
-        logs.push({
-            kind: ValidationKind.Error,
-            reason: ValidationReason.InsufficientTokenBalance,
-            message: "Insufficient token balance to submit transaction.",
-        })
-    } else if (amount > tokenBalance) {
-        logs.push({
-            kind: ValidationKind.Error,
-            reason: ValidationReason.InsufficientTokenBalance,
-            message: "Insufficient token balance to submit transaction.",
-        })
-    }
-
-    // Create a mock tx that calls the substrate extrinsic on pallet-xcm with the same parameters so that we can dry run.
-    const mockTx = createERC20SourceParachainTx(
-        sourceParachain,
-        registry.ethChainId,
-        registry.assetHubParaId,
-        sourceAccountHex,
-        tokenAddress,
-        beneficiaryAccount,
-        amount,
-        fee.totalFeeInDot,
-        messageId,
-        sourceParaId,
-        fee.returnToSenderExecutionFeeDOT,
-        fee.totalFeeInNative !== undefined
-    )
-
-    let sourceDryRunError
-    let assetHubDryRunError
-    if (source.features.hasDryRunApi) {
-        // do the dry run, get the forwarded xcm and dry run that
-        const dryRunSource = await dryRunOnSourceParachain(
-            sourceParachain,
-            registry.assetHubParaId,
-            registry.bridgeHubParaId,
-            mockTx,
-            sourceAccountHex
-        )
-        if (!dryRunSource.success) {
-            logs.push({
-                kind: ValidationKind.Error,
-                reason: ValidationReason.DryRunFailed,
-                message: "Dry run call on source failed.",
-            })
-            sourceDryRunError = dryRunSource.error
-        }
-
-        if (dryRunSource.success && sourceParaId !== registry.assetHubParaId) {
-            if (!dryRunSource.assetHubForwarded) {
-                logs.push({
-                    kind: ValidationKind.Error,
-                    reason: ValidationReason.DryRunFailed,
-                    message: "Dry run call did not provide a forwared xcm.",
-                })
-            } else {
-                const dryRunResultAssetHub = await dryRunAssetHub(
-                    assetHub,
-                    sourceParaId,
-                    registry.bridgeHubParaId,
-                    dryRunSource.assetHubForwarded[1][0]
-                )
-                if (!dryRunResultAssetHub.success) {
-                    logs.push({
-                        kind: ValidationKind.Error,
-                        reason: ValidationReason.DryRunFailed,
-                        message: "Dry run failed on Asset Hub.",
-                    })
-                    assetHubDryRunError = dryRunResultAssetHub.errorMessage
-                }
-            }
-        }
-    } else {
-        logs.push({
-            kind: ValidationKind.Warning,
-            reason: ValidationReason.DryRunApiNotAvailable,
-            message: "Source parachain can not dry run call. Cannot verify success.",
-        })
-        if (sourceParaId !== registry.assetHubParaId) {
-            const dryRunResultAssetHub = await dryRunAssetHub(
-                assetHub,
-                sourceParaId,
-                registry.bridgeHubParaId,
-                buildResultXcmAssetHubERC20TransferFromParachain(
-                    sourceParachain.registry,
-                    registry.ethChainId,
-                    sourceAccountHex,
-                    beneficiaryAccount,
-                    tokenAddress,
-                    "0x0000000000000000000000000000000000000000000000000000000000000000",
-                    amount,
-                    fee.totalFeeInDot,
-                    fee.assetHubExecutionFeeDOT,
-                    sourceParaId,
-                    fee.returnToSenderExecutionFeeDOT,
-                    DOT_LOCATION, // TODO: Support native fee for EVM
-                    DOT_LOCATION
-                )
-            )
-            if (!dryRunResultAssetHub.success) {
-                logs.push({
-                    kind: ValidationKind.Error,
-                    reason: ValidationReason.DryRunFailed,
-                    message: "Dry run failed on Asset Hub.",
-                })
-                assetHubDryRunError = dryRunResultAssetHub.errorMessage
-            }
-        }
-    }
-
-    if (!dotBalance) {
-        logs.push({
-            kind: ValidationKind.Error,
-            reason: ValidationReason.InsufficientDotFee,
-            message: "Could not determine the DOT balance",
-        })
-    } else if (fee.totalFeeInDot > dotBalance) {
-        logs.push({
-            kind: ValidationKind.Error,
-            reason: ValidationReason.InsufficientDotFee,
-            message: "Insufficient DOT balance to submit transaction on the source parachain.",
-        })
-    }
-
-    let feeInfo: FeeInfo | undefined
-    if (logs.length === 0) {
-        const [estimatedGas, feeData] = await Promise.all([
-            sourceEthChain.estimateGas(tx),
-            sourceEthChain.getFeeData(),
-        ])
-        const sourceExecutionFee = (feeData.gasPrice ?? 0n) * estimatedGas
-        if (sourceExecutionFee === 0n) {
-            logs.push({
-                kind: ValidationKind.Error,
-                reason: ValidationReason.FeeEstimationError,
-                message: "Could not get fetch fee details.",
-            })
-        }
-
-        if (sourceExecutionFee > nativeBalance && !nativeBalanceCheckFailed) {
-            logs.push({
-                kind: ValidationKind.Error,
-                reason: ValidationReason.InsufficientNativeFee,
-                message:
-                    "Insufficient native balance to submit transaction on the source parachain.",
-            })
-        }
-        feeInfo = {
-            estimatedGas,
-            feeData,
-            executionFee: sourceExecutionFee,
-            totalTxCost: sourceExecutionFee,
-        }
-    }
-    // Recheck balance after execution fee
-    if (
-        !nativeBalanceCheckFailed &&
-        isNativeBalanceTransfer &&
-        fee.totalFeeInNative &&
-        amount + fee.totalFeeInNative + (feeInfo?.totalTxCost ?? 0n) > tokenBalance
-    ) {
-        logs.push({
-            kind: ValidationKind.Error,
-            reason: ValidationReason.InsufficientTokenBalance,
-            message: "Insufficient token balance to submit transaction.",
-        })
-    }
-    const bridgeStatus = await getOperatingStatus({ gateway, bridgeHub })
-    if (bridgeStatus.toEthereum.outbound !== "Normal") {
-        logs.push({
-            kind: ValidationKind.Error,
-            reason: ValidationReason.BridgeStatusNotOperational,
-            message: "Bridge operations have been paused by onchain governance.",
-        })
-    }
-
-    const success = logs.find((l) => l.kind === ValidationKind.Error) === undefined
-    return {
-        logs,
-        success,
-        data: {
-            bridgeStatus,
-            nativeBalance,
-            dotBalance,
-            feeInfo,
-            tokenBalance,
-            sourceDryRunError,
-            assetHubDryRunError,
-        },
-        transfer,
-    }
-}
-
 export type MessageReceipt = {
     blockNumber: number
     blockHash: string
-    txIndex: number
-    txHash: string
-    success: boolean
-    events: EventRecord[]
-    dispatchError?: any
-    messageId?: string
-}
-
-export type MessageReceiptEvm = {
-    blockNumber: number
-    blockHash: string
-    substrateBlockHash: string
     txIndex: number
     txHash: string
     success: boolean
@@ -1203,59 +751,7 @@ export async function signAndSend(
     return result
 }
 
-export async function getMessageReceipt(
-    sourceParachain: ApiPromise,
-    receipt: TransactionReceipt
-): Promise<MessageReceiptEvm> {
-    const blockHash = await sourceParachain.rpc.chain.getBlockHash(receipt.blockNumber)
-    const events = await (await sourceParachain.at(blockHash)).query.system.events<EventRecord[]>()
-    let success = false
-    let dispatchError: any
-    let messageId: string | undefined
-    const eventTx = events.find(
-        (e) =>
-            sourceParachain.events.ethereum.Executed.is(e.event) &&
-            e.event.data[2].toPrimitive()?.toString().toLowerCase() === receipt.hash.toLowerCase()
-    )
-    if (!(eventTx && eventTx.phase.isApplyExtrinsic)) {
-        throw Error(`Could not find tx hash ${receipt.hash} in block ${receipt.blockNumber}.`)
-    }
-    const matchedEvents: EventRecord[] = events.filter(
-        (e) =>
-            e.phase.isApplyExtrinsic &&
-            e.phase.asApplyExtrinsic.toNumber() === eventTx.phase.asApplyExtrinsic.toNumber()
-    )
-
-    for (const e of matchedEvents) {
-        const data = e.event.data
-        if (sourceParachain.events.system.ExtrinsicFailed.is(e.event)) {
-            dispatchError = data.toHuman(true) as any
-            break
-        } else if (sourceParachain.events.polkadotXcm.Sent.is(e.event)) {
-            success = true
-            const pData = data.toPrimitive()
-            const xcm = (pData as any)[2]
-            messageId = xcm.length > 0 ? xcm[xcm.length - 1].setTopic : (pData as any)[3]
-            break
-        }
-    }
-    if (!messageId) {
-        throw Error(`Not a bridge transfer`)
-    }
-    return {
-        messageId: messageId,
-        blockNumber: receipt.blockNumber,
-        substrateBlockHash: blockHash.toHex(),
-        blockHash: receipt.blockHash,
-        txHash: receipt.hash,
-        txIndex: receipt.index,
-        success: success && receipt.status === 1,
-        dispatchError,
-        events: matchedEvents.map((x) => x.toPrimitive() as any as EventRecord),
-    }
-}
-
-function resolveInputs(registry: AssetRegistry, tokenAddress: string, sourceParaId: number) {
+export function resolveInputs(registry: AssetRegistry, tokenAddress: string, sourceParaId: number) {
     const tokenErcMetadata =
         registry.ethereumChains[registry.ethChainId.toString()].assets[tokenAddress.toLowerCase()]
     if (!tokenErcMetadata) {
@@ -1313,7 +809,7 @@ function createAssetHubTx(
     )
 }
 
-function createERC20SourceParachainTx(
+export function createERC20SourceParachainTx(
     parachain: ApiPromise,
     ethChainId: number,
     assetHubParaId: number,
@@ -1367,7 +863,7 @@ function createERC20SourceParachainTx(
     )
 }
 
-async function dryRunOnSourceParachain(
+export async function dryRunOnSourceParachain(
     source: ApiPromise,
     assetHubParaId: number,
     bridgeHubParaId: number,
@@ -1448,7 +944,7 @@ async function dryRunOnSourceParachain(
     }
 }
 
-async function dryRunAssetHub(
+export async function dryRunAssetHub(
     assetHub: ApiPromise,
     parachainId: number,
     bridgeHubParaId: number,
@@ -1495,7 +991,7 @@ async function dryRunAssetHub(
     }
 }
 
-async function buildMessageId(
+export async function buildMessageId(
     parachain: ApiPromise,
     sourceParaId: number,
     sourceAccountHex: string,
