+++ web/packages/api/src/toPolkadot_v2.ts
@@ -14,14 +14,8 @@ import {
     IGatewayV1 as IGateway,
     IGatewayV1__factory as IGateway__factory,
 } from "@snowbridge/contract-types"
-import {
-    Asset,
-    AssetRegistry,
-    ERC20Metadata,
-    ETHER_TOKEN_ADDRESS,
-    padFeeByPercentage,
-    Parachain,
-} from "./assets_v2"
+import { padFeeByPercentage } from "./assets_v2"
+import { Asset, AssetRegistry, ERC20Metadata, ETHER_TOKEN_ADDRESS, Parachain } from "./types"
 import { getOperatingStatus, OperationStatus } from "./status"
 import { ApiPromise } from "@polkadot/api"
 import {
